Start scrollTo functionality on post opened
I'm done for the day, and recording where I am and what I should do
next.

When a post is opened (thumbnail clicked), the view scrolls to the top
of the post's thumbnail. This is done in the FullPost element.

I need to make it to where there is a suitable behavior for when the
post is made inactive (full post hidden). Right now, it jumps to the top
abruptly -- I need to make it readjust smoothly. Could I use
componentWillUnmount in FullPost?